Dockerfile

+++ create.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-javac -source 1.7 -target 1.7 project/src/*.java
-docker system prune | y
-docker image rm pubsub
-docker image build -t pubsub .
+PROJDIR="/home/ruzan/DS/PubSubSystem/PubSubSystem/src/edu/rit/CSCI652/"
+
+files=`find -name "*.java"`
+javac -d "out" -classpath lib/gson-2.6.2.jar $files
+if [ $? -eq 0 ]
+then
+	echo Compiled..
+	docker image rm pubsub 
+	docker image build -t pubsub .
+fi
